chore: bump react-query 3 to 5 major  update

+++ src/pages/admin/settings/user-details/service/delete-user/delete-user.ts
@@ -1,19 +1,17 @@
-import { useMutation } from "react-query";
+import { useMutation } from "@tanstack/react-query";
 
 import apiClient from "../../../../../../config";
 import { IAPIError } from "../../../../../../types";
 import { API_MUTATION_KEY, APIS_ROUTES } from "../../../../../../utils";
 
 const deleteUser = async (organizationId: string, userId: string) => {
-  await apiClient.delete(
-    `${APIS_ROUTES.DELETE_USER}/${organizationId}/user/${userId}`
-  );
+  await apiClient.delete(`${APIS_ROUTES.DELETE_USER}/${organizationId}/user/${userId}`);
 };
 
 export const useDeleteUser = (organizationId: string) =>
-  useMutation<void, IAPIError, string>(
-    [API_MUTATION_KEY.DELETE_USER],
-    (userId) => deleteUser(organizationId, userId)
-  );
+  useMutation<void, IAPIError, string>({
+    mutationKey: [API_MUTATION_KEY.DELETE_USER],
+    mutationFn: (userId) => deleteUser(organizationId, userId),
+  });
 
 export default useDeleteUser;
